Add permission gate for adding photos to library
